# swagger.yaml
swagger: "2.0"
info:
  description: This is the swagger file that goes with our telemetry api code.
  version: "1.0"
  title: Telemetry API
consumes:
  - application/json
produces:
  - application/json

basePath: /telemetry/api/v1.0/

# Paths supported by the server application
paths:
  /tasks:
    get:
      operationId: appp.get_tasks
      tags:
        - task
      summary: Read the entire list of tasks
      description: Read the list of tasks
      parameters:
        - name: length
          in: query
          type: integer
          description: Number of tasks to get from appp
          required: false
        - name: offset
          in: query
          type: integer
          description: Offset from beginning of list where to start gathering tasks
          required: false
      responses:
        200:
          description: Successfully read tasks list operation
          schema:
            type: array
            items:
              properties:
                APIType:
                  type: string
                AutomationName:
                  type: string
                AutomationServiceAccount:
                  type: string
                CallingAPIKey:
                  type: string
                ExecutionTime:
                  type: integer
                Source:
                  type: string
                Status:
                  type: string
                TargetDevices:
                  type: string
                TicketNo:
                  type: string
                datetimest:
                  type: string
                id:
                  type: integer

    post:
      operationId: appp.create_task
      tags:
        - task
      summary: Create a task and add it to the task list
      description: Create a new task in the task list
      parameters:
        - name: task
          in: body
          description: Task to create
          required: True
          schema:
            type: object
            properties:
              APIType:
                type: string
                description: APIType of task to create
              AutomationName:
                type: string
                description: AutomationName of task to create
              AutomationServiceAccount:
                type: string
                description: AutomationServiceAccount of task to create
              CallingAPIKey:
                type: string
                description: CallingAPIKey of task to create
              ExecutionTime:
                type: integer
                description: ExecutionTime of task to create
              Source:
                type: string
                description: Source of task to create
              Status:
                type: string
                description: Status of task to create
              TargetDevices:
                type: string
                description: TargetDevices of task to create
              TicketNo:
                type: string
                description: TicketNo of task to create
              datetimest:
                type: string
                description: datetimest of task to create
              id:
                type: integer
                description: id of task to create
      responses:
        201:
          description: Successfully created task in list

  /tasks/{CallingAPIKey}:
    get:
      operationId: appp.get_task
      tags:
        - task
      summary: Read one task from the task list
      description: Read one task from the task list
      parameters:
        - name: CallingAPIKey
          in: path
          description: calling API of the task to get from the list
          type: string
          required: True
      responses:
        200:
          description: Successfully read task from task list operation
          schema:
            properties:
              APIType:
                type: string
                description: APIType of task to create
              AutomationName:
                type: string
                description: AutomationName of task to create
              AutomationServiceAccount:
                type: string
                description: AutomationServiceAccount of task to create
              CallingAPIKey:
                type: string
                description: CallingAPIKey of task to create
              ExecutionTime:
                type: integer
                description: ExecutionTime of task to create
              Source:
                type: string
                description: Source of task to create
              Status:
                type: string
                description: Status of task to create
              TargetDevices:
                type: string
                description: TargetDevices of task to create
              TicketNo:
                type: string
                description: TicketNo of task to create
              datetimest:
                type: string
                description: datetimest of task to create
              id:
                type: integer
                description: id of task to create

   